requirements.txt
yfinance
pandas
numpy
matplotlib
plotly
requests
python-dotenv